requirements.txt
web3                      6.8.0
pandas                    2.0.3
numpy                     1.25.1
matplotlib                3.7.2
eth-abi                   4.1.0
eth-account               0.9.0
eth-hash                  0.5.2
eth-keyfile               0.6.1
eth-keys                  0.4.0
eth-rlp                   0.3.0
eth-typing                3.4.0
eth-utils                 2.2.0
etherscan-python          2.1.0
jsonschema                4.19.0
jsonschema-specifications 2023.7.1
urllib3                   2.0.4
web3                      6.8.0
websockets                11.0.3